ENGR00235400 CTS: android.bluetooth.cts.BasicAdapterTest*
Fix the android.bluetooth.cts.BasicAdapterTest* CTS failure on SabreSD.

The cause of the failure is we include the
frameworks/native/data/etc/tablet_core_hardware.xml
which have set the bluetooth feature, so the CTS will run BT's case.

The fix was not include this file and add missed feature in
 required_hardware.xml.

Also remove some redundent hardware features.

Signed-off-by: Zhang Jiejing <jiejing.zhang@freescale.com>